python-chatbot/requirements.txt
langchain>=0.3.0
langchain-ollama>=0.2.0
langserve[all]>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
sse-starlette>=2.1.0
ollama>=0.4.0
